backend/pyproject.toml
[project]
name = "relativity-backend"
version = "0.1.0"
description = "RAG-powered legal AI platform backend"
requires-python = ">=3.10"
dependencies = [
  # The web framework — handles HTTP routing, request/response, validation
  "fastapi>=0.115.0",
  # The server that actually runs FastAPI (like Vite's dev server, but for Python)
  "uvicorn[standard]>=0.32.0",
  # ORM — lets us write Python classes instead of raw SQL to define/query tables
  "sqlalchemy>=2.0.0",
  # Migration tool — tracks and applies database schema changes over time
  "alembic>=1.14.0",
  # PostgreSQL driver — the bridge between Python and PostgreSQL
  "psycopg2-binary>=2.9.0",
  # Reads .env files into typed Python settings objects
  "pydantic-settings>=2.7.0",
  # Required for FastAPI to handle file uploads (multipart/form-data)
  "python-multipart>=0.0.12",
  "pdfminer-six>=20221105",
  "python-docx>=1.1.0",
]

[dependency-groups]
dev = [
  # Test runner (like Vitest but for Python)
  "pytest>=8.0.0",
  # Adds async test support — needed because FastAPI handlers can be async
  "pytest-asyncio>=0.24.0",
  # HTTP client used in tests to call our API endpoints
  "httpx>=0.28.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
